Move data structures to src directory
