--- Set.java
@@ -1,19 +1,0 @@
-
-// Set<E> is a genaric interface for the set (or dictionary) ADT.
-
-public interface Set<E> {
-    // TODO: comment
-    public void add(E value);
-
-    // TODO: comment
-    public void remove(E value);
-
-    // TODO: comment
-    public boolean contains(E value);
-
-    // TODO: comment
-    public boolean isEmpty();
-
-    // TODO: comment
-    public int size();
-}
