+++ src/main/java/com/github/hubbards/data/structures/UnderflowException.java
@@ -1,0 +1,26 @@
+package com.github.hubbards.data.structures;
+
+/**
+ * This class represents a runtime exception for operation on an empty
+ * collection, e.g., list, queue, or heap.
+ * 
+ * @author Spencer Hubbard
+ */
+@SuppressWarnings("serial")
+public class UnderflowException extends RuntimeException {
+    /**
+     * Constructs a new underflow exception.
+     */
+    public UnderflowException() {
+        super();
+    }
+    
+    /**
+     * Constructs a new underflow exception with a given message.
+     * 
+     * @param message the given message
+     */
+    public UnderflowException(String message) {
+        super(message);
+    }
+}
