+++ src/test/java/com/github/hubbards/data/structures/LinearProbingHashTableTest.java
@@ -1,0 +1,13 @@
+package com.github.hubbards.data.structures;
+
+/**
+ * TODO document class
+ */
+public class LinearProbingHashTableTest extends HashTableTest<LinearProbingHashTable<String>> {
+    @Override
+    protected LinearProbingHashTable<String> createHashTable() {
+        return new LinearProbingHashTable<String>();
+    }
+
+    // TODO write implementation specific tests
+}
