+++ src/test/java/com/github/hubbards/data/structures/LinkedListTest.java
@@ -1,0 +1,11 @@
+package com.github.hubbards.data.structures;
+
+/**
+ * TODO document class
+ */
+public class LinkedListTest extends ListTest<LinkedList<Integer>> {
+    @Override
+    protected LinkedList<Integer> createList() {
+        return new LinkedList<Integer>();
+    }
+}
