+++ src/test/java/com/github/hubbards/data/structures/ListTest.java
@@ -1,0 +1,72 @@
+package com.github.hubbards.data.structures;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * TODO document class
+ */
+public abstract class ListTest<T extends List<Integer>> {
+    private T list;
+
+    protected abstract T createList();
+
+    @Before
+    public void setUp() {
+        list = createList();
+    }
+
+    @Test
+    public void testIsEmpty1() {
+        assertTrue(list.isEmpty());
+    }
+
+    @Test
+    public void testIsEmpty2() {
+        list.add(1);
+
+        assertFalse(list.isEmpty());
+    }
+
+    @Test
+    public void testIsEmpty3() {
+        list.add(1);
+        list.add(2);
+        list.remove(1);
+
+        assertFalse(list.isEmpty());
+    }
+
+    @Test
+    public void testIsEmpty4() {
+        list.add(1);
+        list.remove(0);
+
+        assertTrue(list.isEmpty());
+    }
+
+    @Test
+    public void testSize1() {
+        assertEquals(0, list.size());
+    }
+
+    @Test
+    public void testSize2() {
+        list.add(1);
+
+        assertEquals(1, list.size());
+    }
+
+    @Test
+    public void testSize3() {
+        list.add(1);
+        list.add(2);
+        list.remove(1);
+
+        assertEquals(1, list.size());
+    }
+
+    // TODO: write more tests
+}
