+++ src/test/java/com/github/hubbards/data/structures/SeparateChainingHashTableTest.java
@@ -1,0 +1,13 @@
+package com.github.hubbards.data.structures;
+
+/**
+ * TODO document class
+ */
+public class SeparateChainingHashTableTest extends HashTableTest<SeparateChainingHashTable<String>> {
+    @Override
+    protected SeparateChainingHashTable<String> createHashTable() {
+        return new SeparateChainingHashTable<String>();
+    }
+
+    // TODO write implementation specific tests
+}
